tripza/requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-crispy-forms==1.8.1
django-elasticsearch-dsl==7.3
django-environ==0.9.0
django-filter==22.1
django-smart-selects==1.6.0
django-tables2==2.5.2
django-tinymce4-lite==1.8.0
django-utils-six==2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
environ==1.0
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
jsmin==3.0.1
Markdown==3.4.1
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
packaging==23.1
pandas==1.5.3
Pillow==9.4.0
promise==2.3
protobuf==3.20.3
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
text-unidecode==1.3
threadpoolctl==3.1.0
tqdm==4.65.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
